refactor(Project): move Transformation, Composite base classes to top-level base.py
