--- src/fold/all_types.py
@@ -1,6 +1,0 @@
-from typing import List
-
-import pandas as pd
-
-TrainedPipelines = List[pd.Series]
-OutOfSamplePredictions = pd.DataFrame
