+++ src/fold/composites/__init__.py
@@ -1,4 +1,4 @@
-from .base import Composite
+from ..base import Composite
 from .columns import PerColumnEnsemble, PerColumnTransform, SkipNA
 from .ensemble import Ensemble
 from .hybrid import Hybrid
