--- src/fold/composites/base.py
@@ -1,77 +1,0 @@
-from abc import ABC, abstractmethod
-from dataclasses import dataclass
-from typing import Callable, List, Optional, Tuple, TypeVar
-
-import pandas as pd
-
-from ..transformations.base import Pipelines
-
-T = TypeVar("T", Optional[pd.Series], pd.Series)
-
-
-class Composite(ABC):
-    """
-    A Composite contains other transformations.
-    """
-
-    @dataclass
-    class Properties:
-        primary_requires_predictions: bool = (
-            False  # Primary pipeline need output from a model
-        )
-        primary_only_single_pipeline: bool = (
-            False  # There should be a single primary pipeline.
-        )
-        secondary_requires_predictions: bool = (
-            False  # Secondary pipeline need output from a model
-        )
-        secondary_only_single_pipeline: bool = (
-            False  # There should be a single secondary pipeline.
-        )
-
-    properties: Properties
-
-    @abstractmethod
-    def get_child_transformations_primary(self) -> Pipelines:
-        raise NotImplementedError
-
-    def get_child_transformations_secondary(
-        self,
-    ) -> Optional[Pipelines]:
-        return None
-
-    @abstractmethod
-    def postprocess_result_primary(
-        self, results: List[pd.DataFrame], y: Optional[pd.Series]
-    ) -> pd.DataFrame:
-        raise NotImplementedError
-
-    def postprocess_result_secondary(
-        self,
-        primary_results: List[pd.DataFrame],
-        secondary_results: List[pd.DataFrame],
-        y: Optional[pd.Series],
-    ) -> pd.DataFrame:
-        raise NotImplementedError
-
-    @abstractmethod
-    def clone(self, clone_child_transformations: Callable) -> "Composite":
-        raise NotImplementedError
-
-    def before_fit(self, X: pd.DataFrame) -> None:
-        pass
-
-    def preprocess_primary(
-        self, X: pd.DataFrame, index: int, y: T, fit: bool
-    ) -> Tuple[pd.DataFrame, T]:
-        return X, y
-
-    def preprocess_secondary(
-        self,
-        X: pd.DataFrame,
-        y: T,
-        results_primary: List[pd.DataFrame],
-        index: int,
-        fit: bool,
-    ) -> Tuple[pd.DataFrame, T]:
-        return X, y
