+++ src/fold/loop/convenience.py
@@ -8,8 +8,7 @@ from fold.models.base import Model
 from fold.models.sklearn import SKLearnClassifier, SKLearnPipeline, SKLearnRegressor
 from fold.transformations.sklearn import SKLearnFeatureSelector, SKLearnTransformation
 
-from ..composites.base import Composite
-from ..transformations.base import BlocksOrWrappable, Transformation, Transformations
+from ..base import BlocksOrWrappable, Composite, Transformation, Transformations
 from ..transformations.function import WrapFunction
