+++ src/fold/models/random.py
@@ -4,7 +4,7 @@ from typing import List, Union
 import numpy as np
 import pandas as pd
 
-from ..transformations.base import Transformation, fit_noop
+from ..base import Transformation, fit_noop
 from .base import Model
